Don't remove work_dir if we're running locally, add migration file.

+++ workers/data_refinery_workers/processors/utils.py
@@ -24,11 +24,12 @@ from data_refinery_common.models import (
     Sample,
 )
 from data_refinery_workers._version import __version__
-from data_refinery_common.utils import get_instance_id, get_env_variable
+from data_refinery_common.utils import get_instance_id, get_env_variable, get_env_variable_gracefully
 
 
 logger = get_and_configure_logger(__name__)
 S3_BUCKET_NAME = get_env_variable("S3_BUCKET_NAME", "data-refinery")
+RUNNING_IN_CLOUD = get_env_variable_gracefully("RUNNING_IN_CLOUD", "False")
 DIRNAME = os.path.dirname(os.path.abspath(__file__))
 
 
@@ -158,7 +159,7 @@ def end_job(job_context: Dict, abort=False):
         if len(pipeline.steps):
             pipeline.save()
 
-    if "work_dir" in job_context:
+    if "work_dir" in job_context and RUNNING_IN_CLOUD == "True":
         shutil.rmtree(job_context["work_dir"])
 
     job.success = success
